added mobile layout, hamburger menu, refactor initial layout, refactor components to mobile layout
