+++ src/app/Layouts/components/footer/version-info.tsx
@@ -1,0 +1,8 @@
+import { Text } from '@chakra-ui/react';
+import { T } from "../../../shared/components";
+
+export const VersionInfo = () => (
+    <Text fontSize="xs">
+        <T dictKey="devVersion" />
+    </Text>
+);
