+++ src/app/Layouts/components/menu/index.ts
@@ -1,0 +1,1 @@
+export { renderMenu } from './render-menu';
