+++ src/app/Layouts/components/mobile-header/index.ts
@@ -1,0 +1,1 @@
+export { MobileHeader } from './mobile-header';
