+++ src/app/Layouts/current-layout.tsx
@@ -1,0 +1,7 @@
+import { getCurrentLayout } from "../shared/utils";
+
+export const CurrentLayout = () => {
+    const LayoutType = getCurrentLayout();
+
+    return <LayoutType  />
+};
