+++ src/app/Layouts/index.ts
@@ -1,0 +1,1 @@
+export { CurrentLayout } from './current-layout';
