+++ src/app/Layouts/layout-types/types.ts
@@ -1,0 +1,5 @@
+export interface ILayoutType {
+    headerHeightPx: string;
+    footerHeightPx: string;
+    navWidthPx: string;
+}
