+++ src/app/features/map2d/containers/map2d.tsx
@@ -16,7 +16,7 @@ export const Map2d = () => {
     const setNoradId = (id: string): void => setSelectedNoradId(id);
 
     return (    
-        <Box h="100%" pr={MAP_PADDING} pl={MAP_PADDING} pt={MAP_PADDING}>
+        <Box h="100%" pr={MAP_PADDING} pl={MAP_PADDING} pt={MAP_PADDING} pb={MAP_PADDING}>
             <MapAllSats 
                 setNoradId={setNoradId} 
                 setOpenDrawer={openDrawer}
