+++ src/app/shared/utils/get-current-layout.ts
@@ -1,0 +1,11 @@
+import { FullWidthLayout, MobileLayout } from "../../Layouts/layout-types";
+
+export const getCurrentLayout = () => {
+    const mobileBreakpoint = 768;
+
+    if (window.innerWidth <= mobileBreakpoint) {
+        return MobileLayout;
+    } else {
+        return FullWidthLayout;
+    }
+}
